// agent-front/src/components/layout/types.ts
import { type LinkProps } from '@tanstack/react-router'
import { type IconName } from 'lucide-react/dynamic'

type NavigationIcon = React.ElementType | IconName

type User = {
  name: string
  email: string
  avatar: string
}

type Team = {
  name: string
  logo: React.ElementType
  plan: string
}

type BaseNavItem = {
  title: string
  badge?: string
  icon?: NavigationIcon
}

type NavLink = BaseNavItem & {
  url: LinkProps['to'] | (string & {})
  items?: never
}

type NavCollapsible = BaseNavItem & {
  items: (BaseNavItem & { url: LinkProps['to'] | (string & {}) })[]
  url?: never
}

type NavItem = NavCollapsible | NavLink

type NavGroup = {
  title: string
  items: NavItem[]
}

type SidebarData = {
  user: User
  teams: Team[]
  navGroups: NavGroup[]
}

export type {
  NavigationIcon,
  SidebarData,
  NavGroup,
  NavItem,
  NavCollapsible,
  NavLink,
}